Update AssemblyInfo and handle single class files
AssemblyInfo's AssemblyInformationalVersionAttribute has been updated with the correct version. On another note, a condition has been added to handle files that contain only one class. Instead of going through the splitting process, the program will now return and output a message stating that the file contains a single class and was not split.